Update kubernetes-deployment.md
Example forward command works for helm, but not pure manifest (even from helm). 

Maybe not worth including, but the page is already mentioning two ways to install (helm and not-helm). Personally, I avoid helm at all cost, and per this page, so do others. Be nice to have complete steps (even if easy to reverse engineer and fix).